implement seek partially for files opened on the http fs
we only allow seeking to the start (when already at the start) and
to the end (to let callers determine file size). we only do this
so net/http's ServeContent will work.

we could attempt to implement full seek for non-compressed files
(which is what i use always anyways). maybe later.