work better on windows, easier-to-use example
on windows, if opening os.Args[0] fails, we try the same but with
".exe" appended. if you launch a program from the command-line, you
don't have to type the ".exe".  before this change, that meant we
couldn't find the binary, or the zip file it contains.

a full example has been added. this is the example from the package
docs.

also slight tweak to license.